requirements.txt
Django>=2.2,<3
mysqlclient
djangorestframework
pymysql
netaddr
pyllist
django-cors-headers
google-auth
